+++ CHANGES.md
@@ -56,6 +56,7 @@ Changes in 12.5
 - JS remove inline code from program in Rollover.php & Rollover.js
 - JS remove inline code from `_makePaymentsCommentsInput()` in Student_Billing/functions.inc.php & MakePaymentsCommentsInput.js
 - JS remove inline code from Addresses & Contacts tab in Address.inc.php & Address.js
+- JS remove inline code from General Info tab in General_Info.inc.php & CreateStudentAccount.js
 
 Changes in 12.4.3
 -----------------